Arborescence des fichiers + ajout style games selection
