--- src/app/api/games/select-games/route.ts
@@ -1,6 +1,0 @@
-import { NextRequest } from "next/server";
-
-export async function POST(request: NextRequest) {
-    // TODO
-    // Implémenter la logique pour gérer la sélection d'un ou des jeux
-}
